cherry-picked from custom_tabs: Utils modifications
