+++ app/src/hnqis/java/org/eyeseetea/malariacare/utils/Utils.java
@@ -31,7 +31,6 @@ import android.view.View;
 import android.widget.Button;
 import android.widget.ImageView;
 import android.widget.TextView;
-
 import org.eyeseetea.malariacare.BuildConfig;
 import org.eyeseetea.malariacare.R;
